+++ app/src/main/java/com/github/sdpsharelook/authorization/User.kt
@@ -1,0 +1,9 @@
+package com.github.sdpsharelook.authorization
+
+data class User(
+    val uid: String,
+    val email: String,
+    val displayName: String = "anonymous",
+    val isAnonymous: Boolean = true
+)
+
