+++ app/src/test/java/com/github/sdpsharelook/authorization/TestUserConstants.kt
@@ -1,0 +1,8 @@
+package com.github.sdpsharelook.authorization
+
+object TestUserConstants {
+    const val TEST_USER_EMAIL = "testuser@gmail.com"
+    const val TEST_USER_UID = "DSKjhf23dskjfhalfaJFSD"
+    const val TEST_USER_PASS = "123456"
+    const val TEST_USER_PASS2 = "Abcdef1!"
+}
